updates: Fixed navigation bar so that it now fixed at the top of the page, even when scrolling.
Also reogranized the div containers so that all of the main content is
in one large div.

Css changes are within the page at the top of the code, but it would be
better to make the alterations in a separate file.

If I edit pre-exiting styles in a separate file, how do I make my custom
styles priority?